Fix the quad rasteriser in 64 bit builds.
git-svn-id: https://rpg.hamsterrepublic.com/source/wip@7993 7d344553-34f0-0310-a9b1-970ce8f1c3a2

+++ matrixMath.h
@@ -17,7 +17,7 @@ struct float3x3 {
          _31, _32, _33;
 };
 struct SURFACE_RECT {
-   long left, top, right, bottom;
+   int left, top, right, bottom;
 };
 
 extern "C" {
